requirements.txt
django>1.8.0
# mysqlclient>1.3.6
celery>3.1
djangorestframework>3.6
markdown
django-filter
django-celery>3.1

